+++ crud/common/schema.lua
@@ -197,19 +197,14 @@ function schema.truncate_row_trailing_fields(tuple, field_names)
     return tuple
 end
 
--- schema.wrap_box_space_func_result pcalls some box.space function
--- and returns its result as a table
--- `{res = ..., err = ..., space_schema_hash = ...}`
--- space_schema_hash is computed if function failed and
--- `add_space_schema_hash` is true
-function schema.wrap_box_space_func_result(space, func_name, args, opts)
-    dev_checks('table', 'string', 'table', 'table')
+function schema.wrap_func_result(space, func, args, opts)
+    dev_checks('table', 'function', 'table', 'table')
 
     local result = {}
 
     opts = opts or {}
 
-    local ok, func_res = pcall(space[func_name], space, unpack(args))
+    local ok, func_res = pcall(func, unpack(args))
     if not ok then
         result.err = func_res
         if opts.add_space_schema_hash then
@@ -220,6 +215,20 @@ function schema.wrap_box_space_func_result(space, func_name, args, opts)
     end
 
     return result
+end
+
+-- schema.wrap_box_space_func_result pcalls some box.space function
+-- and returns its result as a table
+-- `{res = ..., err = ..., space_schema_hash = ...}`
+-- space_schema_hash is computed if function failed and
+-- `add_space_schema_hash` is true
+function schema.wrap_box_space_func_result(space, box_space_func_name, box_space_func_args, opts)
+    dev_checks('table', 'string', 'table', 'table')
+    local function func(space, box_space_func_name, box_space_func_args)
+        return space[box_space_func_name](space, unpack(box_space_func_args))
+    end
+
+    return schema.wrap_func_result(space, func, {space, box_space_func_name, box_space_func_args}, opts)
 end
 
 -- schema.result_needs_reload checks that schema reload can